last2days

+++ routes/comments.routes.js
@@ -9,6 +9,10 @@ router.get("/getCommentByFest/:fest_id", (req, res, next) => {
     Fest
         .findById(fest_id)
         .populate('owner comments')
+        .populate({
+            path: 'comments',
+            populate: { path: 'owner' }
+        })
         .then(response => {
             res.json(response.comments)
         })
@@ -25,11 +29,9 @@ router.post("/newComment/:fest_id", verifyToken, (req, res, next) => {
     Comment
         .create({ text, createdAt, owner })
         .then(response => {
-            Fest
-                .findByIdAndUpdate(fest_id, { $addToSet: { comments: response._id } })
-                .then((fest) => res.json(fest))
-                .catch(err => next(err))
+            return Fest.findByIdAndUpdate(fest_id, { $addToSet: { comments: response._id } })
         })
+        .then((fest) => res.json(fest))
         .catch(err => next(err))
 })
 
@@ -39,19 +41,23 @@ router.put('/edit-comment/:comment_id', (req, res, next) => {
     const { text, createdAt } = req.body
 
     Comment
-
         .findByIdAndUpdate(comment_id, { text, createdAt })
         .then(response => res.json(response))
         .catch(err => next(err))
 })
 
-router.delete('/delete-comment/:comment_id', (req, res, next) => {
+router.delete('/delete-comment/:comment_id', verifyToken, (req, res, next) => {
 
-    const { comment_id } = req.params
+    const { comment_id, fest_id } = req.params
+    const { _id: owner } = req.payload
 
     Comment
-        .findByIdAndDelete(comment_id)
-        .then(response => res.json(response))
+        .findByIdAndDelete(comment_id, { owner })
+        .then(response => {
+            console.log('----', response)
+            return Fest.findByIdAndUpdate(fest_id, { $pull: { comments: response._id } })
+        })
+        .then(fest => res.json(fest))
         .catch(err => next(err))
 })
